fix: disable re-org checks for ElectrumX.
ElectrumX was crashing after I instructed simnet to mine 100 blocks. It seems like it was an issue with the async block processing logic in Electrum. Not sure why I didn't encounter this while using bitcoind and electrumx, but this appears to fix it for now. We can safely assume no re-orgs for local testing purposes.